refactor: Add ping handler to serve a simple "OK" response
This commit adds a new handler function called "ping" to the handlers.go file. The handler responds with a simple "OK" message when a GET request is made to the "/" endpoint. This handler is useful for checking if the server is running properly. The commit also includes a new test function in the handlers_test.go file to test the functionality of the ping handler.

Ref: #123